backend/requirements.txt
marshmallow>=3
flask
flask-restful
flask-migrate
flask-sqlalchemy
flask-marshmallow
flask-jwt-extended
marshmallow-sqlalchemy
python-dotenv
passlib
apispec[yaml]
apispec-webframeworks
tox
gunicorn
colorama
sqlalchemy_utils
Babel
requests